+++ src/main.rs
@@ -1,4 +1,5 @@
 mod monster;
+mod parts;
 use crate::monster::Monster;
 use crate::monster::{head::Head, body::Body, arms::Arms, legs::Legs};
 use nannou::prelude::*;
