+++ src/monster/legs.rs
@@ -1,3 +1,4 @@
+use crate::monster::Rawr;
 use nannou::color::encoding::Srgb;
 use nannou::color::rgb::Rgb;
 use nannou::prelude::*;
